pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "okxtrendbot"
version = "0.1.0"
description = "Independent OKX BTC quantitative trend-following bot"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
okxtrendbot = "okxtrendbot.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

